requirements.txt
Django==1.5.4
dj-database-url==0.2.2
gevent==0.13.8
greenlet==0.4.1
gunicorn==18.0
psycopg2==2.5.1
wsgiref==0.1.2
pymongo==2.6.2
jaro-winkler==1.0.2
Yelpy
